updated: changed package namespace to net.azzerial in the api and playground modules
