--- api/src/main/java/com/github/azzerial/slash/internal/ComponentCallback.java
@@ -1,42 +1,0 @@
-/*
- * Copyright 2021 Robin Mercier
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.github.azzerial.slash.internal;
-
-import java.lang.reflect.Method;
-
-public final class ComponentCallback {
-
-    private final Object obj;
-    private final Method method;
-
-    /* Constructors */
-
-    ComponentCallback(Object obj, Method method) {
-        this.obj = obj;
-        this.method = method;
-    }
-
-    /* Getters & Setters */
-
-    public Object getObjectInstance() {
-        return obj;
-    }
-
-    public Method getMethod() {
-        return method;
-    }
-}
